% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{add_annotation}
\alias{add_annotation}
\title{Apply annotation}
\usage{
add_annotation(
  g,
  x,
  y,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  axis_text_x_angle = 45,
  axis_text_x_hjust = 1,
  axis_text_x_vjust = 1,
  legend_position = "bottom",
  n_breaks_y = 20,
  ...
)
}
\description{
Apply annotation
}
