% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{add_axis}
\alias{add_axis}
\title{Apply axis}
\usage{
add_axis(
  g,
  log_x = F,
  log_y = F,
  xlim = NULL,
  ylim = NULL,
  units_x = NULL,
  units_y = NULL,
  ...
)
}
\description{
Apply axis
}
