% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R
\name{add_boxplot}
\alias{add_boxplot}
\title{Add Boxplot}
\usage{
add_boxplot(
  g,
  mapping,
  dodge = 0.5,
  outlier.shape = NA,
  coef = 0,
  alpha = 1,
  width = 0.2,
  color = "black",
  ...
)
}
\description{
Add Boxplot
}
