% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R
\name{add_facets}
\alias{add_facets}
\title{Add facet_grid}
\usage{
add_facets(
  g,
  type = "grid",
  column = NULL,
  row = NULL,
  scales = "free",
  labeller = "label_both",
  ...
)
}
\description{
Add facet_grid
}
