% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R
\name{add_geom}
\alias{add_geom}
\title{Add generic}
\usage{
add_geom(
  geom,
  g,
  mapping = list(),
  dodge = 0.5,
  color = "black",
  cut_f = ggplot2::cut_interval,
  cut_args = list(n = 5),
  ...
)
}
\description{
Add generic
}
