% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_plots.R
\name{add_inline_forestplot}
\alias{add_inline_forestplot}
\title{MISSING_TITLE}
\usage{
add_inline_forestplot(
  x,
  xmin,
  xmax,
  xlim = NULL,
  xintercept = NULL,
  xlab = NULL,
  ylab = NULL,
  hide_y_axis = TRUE,
  hide_x_axis = TRUE,
  hide_legend = TRUE,
  height = 0.3,
  width = 3,
  scaling = 0.8,
  shape = 15,
  size = 4.5,
  alpha = 0.8,
  breaks = seq(-10, 10, 0.5),
  theme = ggplot2::theme_void
)
}
\description{
MISSING_TITLE
}
