% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_plots.R
\name{add_inline_plot}
\alias{add_inline_plot}
\title{MISSING_TITLE}
\usage{
add_inline_plot(
  x,
  y = NULL,
  mapping = list(x = "x", y = "y"),
  style = NULL,
  xlim = NULL,
  ylim = NULL,
  height = 0.8,
  width = 8,
  scaling = 1,
  theme = ggplot2::theme_void,
  coord_flip = FALSE,
  hide_title = T,
  hide_legend = T,
  hide_x_axis = T,
  hide_y_axis = T,
  fill = picasso::roche_colors("lightblue"),
  expansion_y = 0,
  add_download_button = FALSE,
  ...
)
}
\description{
MISSING_TITLE
}
