% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{add_x_axis}
\alias{add_x_axis}
\title{Add x-axis to plot (compatible with picasso::add_x_axis)}
\usage{
add_x_axis(legend_position = "none")
}
\arguments{
\item{legend_position}{String. Legend position. Default "none"}
}
\value{
ggplot2 theme element
}
\description{
Add x-axis to plot (compatible with picasso::add_x_axis)
}
