% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R
\name{aes}
\alias{aes}
\title{Create aes from list
Setup ggplot
This was difficult, fist store parameters in list,
Convert to symbols
drop the onses which are null, call aes_ function (CAVE: ecex)
also think of other places where params is used, e.g. params$id}
\usage{
aes(aes)
}
\description{
Create aes from list
Setup ggplot
This was difficult, fist store parameters in list,
Convert to symbols
drop the onses which are null, call aes_ function (CAVE: ecex)
also think of other places where params is used, e.g. params$id
}
