% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{apply_annotation}
\alias{apply_annotation}
\title{Apply annotation}
\usage{
apply_annotation(
  g,
  x,
  y,
  title,
  xlab,
  ylab,
  axis_text_x,
  axis_text_x_angle,
  axis_text_x_hjust,
  axis_text_x_vjust,
  legend_position,
  n_breaks_y = 20
)
}
\description{
Apply annotation
}
