% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traformations.R
\name{calc_frequency}
\alias{calc_frequency}
\title{Calculate frequency transformation for plotting}
\usage{
calc_frequency(data, mapping)
}
\arguments{
\item{data}{Data frame containing the variables}

\item{mapping}{Aesthetic mapping list containing variable names}
}
\value{
List containing transformed data and updated mapping
}
\description{
Calculates relative frequencies for categorical data transformation in plots
}
