% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{calc_summary_categorical}
\alias{calc_summary_categorical}
\title{MISSING_TITLE}
\usage{
calc_summary_categorical(
  data,
  column,
  calculations = list(N. = nightowl::n, Freq = nightowl::frequencies, Bar = function(x)
    {     nightowl::frequencies(x, output = "barplot") }),
  unnest = TRUE,
  names_sep = NULL
)
}
\value{

}
\description{
MISSING_TITLE
}
