% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{calc_summary_numeric}
\alias{calc_summary_numeric}
\title{MISSING_TITLE}
\usage{
calc_summary_numeric(
  data,
  column,
  calculations = list(N = nightowl::n, Median = median, Mean = ggplot2::mean_cl_boot),
  unnest = TRUE
)
}
\value{

}
\description{
MISSING_TITLE
}
