% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{count}
\alias{count}
\title{Calculate Count Transformation}
\usage{
count(data, mapping)
}
\arguments{
\item{data}{Data frame containing the data to count}

\item{mapping}{List specifying variable mappings}
}
\value{
List containing counted data with n column and updated mapping
}
\description{
Counts occurrences of combinations in grouped data
}
