% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{Coxph}
\alias{Coxph}
\title{R6 Class}
\description{
R6 Class

R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Coxph-new}{\code{Coxph$new()}}
\item \href{#method-Coxph-print}{\code{Coxph$print()}}
\item \href{#method-Coxph-check_time}{\code{Coxph$check_time()}}
\item \href{#method-Coxph-check_event}{\code{Coxph$check_event()}}
\item \href{#method-Coxph-check_treatment}{\code{Coxph$check_treatment()}}
\item \href{#method-Coxph-get_variables}{\code{Coxph$get_variables()}}
\item \href{#method-Coxph-check_variables}{\code{Coxph$check_variables()}}
\item \href{#method-Coxph-set_data}{\code{Coxph$set_data()}}
\item \href{#method-Coxph-check_data}{\code{Coxph$check_data()}}
\item \href{#method-Coxph-set_reference}{\code{Coxph$set_reference()}}
\item \href{#method-Coxph-get_reference}{\code{Coxph$get_reference()}}
\item \href{#method-Coxph-set_formula}{\code{Coxph$set_formula()}}
\item \href{#method-Coxph-check_formula}{\code{Coxph$check_formula()}}
\item \href{#method-Coxph-fit}{\code{Coxph$fit()}}
\item \href{#method-Coxph-clone}{\code{Coxph$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-new"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-print"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-check_time"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-check_time}{}}}
\subsection{Method \code{check_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$check_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-check_event"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-check_event}{}}}
\subsection{Method \code{check_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$check_event()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-check_treatment"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-check_treatment}{}}}
\subsection{Method \code{check_treatment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$check_treatment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-get_variables"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-get_variables}{}}}
\subsection{Method \code{get_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$get_variables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-check_variables"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-check_variables}{}}}
\subsection{Method \code{check_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$check_variables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-set_data}{}}}
\subsection{Method \code{set_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$set_data(data = self$data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-check_data"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-check_data}{}}}
\subsection{Method \code{check_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$check_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-set_reference"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-set_reference}{}}}
\subsection{Method \code{set_reference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$set_reference(data = self$data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-get_reference"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-get_reference}{}}}
\subsection{Method \code{get_reference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$get_reference()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-set_formula"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-set_formula}{}}}
\subsection{Method \code{set_formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$set_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-check_formula"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-check_formula}{}}}
\subsection{Method \code{check_formula()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$check_formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Coxph-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Coxph-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coxph$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
