% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.r
\name{create_formula}
\alias{create_formula}
\title{MISSING_TITLE}
\usage{
create_formula(
  data,
  response,
  treatment,
  covariates = NULL,
  strata = NULL,
  interactions = NULL,
  random_effects = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{treatment}{\code{string} the treatment variable}

\item{covariates}{\verb{character vector} of covariate variables}

\item{strata}{\verb{character vector} of strata variables}

\item{interactions}{\code{list} of interaction terms}

\item{random_effects}{\verb{character vector} of random effects variables}

\item{time}{\code{string} the time variable}

\item{event}{\code{string} the event variable}
}
\description{
MISSING_TITLE
}
