% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{data_error}
\alias{data_error}
\title{Data error for data-related issues}
\usage{
data_error(message, data_info = NULL, call = NULL)
}
\arguments{
\item{message}{Error message}

\item{data_info}{Information about the problematic data}

\item{call}{The call that generated the error}
}
\value{
An error object of class 'data_error'
}
\description{
Data error for data-related issues
}
\keyword{internal}
