% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{data_summary}
\alias{data_summary}
\title{MISSING_TITLE}
\usage{
data_summary(
  data,
  column,
  group_by = NULL,
  .mean = ggplot2::mean_cl_boot,
  .summarise_numeric = nightowl::summarise_numeric,
  .summarise_categorical = nightowl::summarise_categorical,
  calc_p = TRUE,
  show_p = TRUE,
  add_caption = TRUE,
  wrap_header = TRUE,
  output = "raw",
  labels = NULL,
  keep_y = FALSE,
  .range = NULL,
  arrange_by = NULL,
  options_reactables = list(defaultPageSize = 30),
  ...
)
}
\description{
MISSING_TITLE
}
