% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{.NightowlOptions}
\alias{.NightowlOptions}
\title{R6 Class}
\description{
R6 Class

R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NightowlOptions-new}{\code{.NightowlOptions$new()}}
\item \href{#method-NightowlOptions-set_colors}{\code{.NightowlOptions$set_colors()}}
\item \href{#method-NightowlOptions-get_colors}{\code{.NightowlOptions$get_colors()}}
\item \href{#method-NightowlOptions-set_missing_color}{\code{.NightowlOptions$set_missing_color()}}
\item \href{#method-NightowlOptions-get_missing_color}{\code{.NightowlOptions$get_missing_color()}}
\item \href{#method-NightowlOptions-set_header_width}{\code{.NightowlOptions$set_header_width()}}
\item \href{#method-NightowlOptions-get_header_width}{\code{.NightowlOptions$get_header_width()}}
\item \href{#method-NightowlOptions-clone}{\code{.NightowlOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NightowlOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-NightowlOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.NightowlOptions$new(colors = NULL, missingColor = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NightowlOptions-set_colors"></a>}}
\if{latex}{\out{\hypertarget{method-NightowlOptions-set_colors}{}}}
\subsection{Method \code{set_colors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.NightowlOptions$set_colors(colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NightowlOptions-get_colors"></a>}}
\if{latex}{\out{\hypertarget{method-NightowlOptions-get_colors}{}}}
\subsection{Method \code{get_colors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.NightowlOptions$get_colors(n = NULL, missing = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NightowlOptions-set_missing_color"></a>}}
\if{latex}{\out{\hypertarget{method-NightowlOptions-set_missing_color}{}}}
\subsection{Method \code{set_missing_color()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.NightowlOptions$set_missing_color(color)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NightowlOptions-get_missing_color"></a>}}
\if{latex}{\out{\hypertarget{method-NightowlOptions-get_missing_color}{}}}
\subsection{Method \code{get_missing_color()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.NightowlOptions$get_missing_color()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NightowlOptions-set_header_width"></a>}}
\if{latex}{\out{\hypertarget{method-NightowlOptions-set_header_width}{}}}
\subsection{Method \code{set_header_width()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.NightowlOptions$set_header_width(width)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NightowlOptions-get_header_width"></a>}}
\if{latex}{\out{\hypertarget{method-NightowlOptions-get_header_width}{}}}
\subsection{Method \code{get_header_width()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.NightowlOptions$get_header_width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NightowlOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NightowlOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.NightowlOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
