% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{frequencies}
\alias{frequencies}
\title{MISSING_TITLE}
\usage{
frequencies(
  x,
  output = "print",
  digits = 1,
  str_width = NightowlOptions$get_header_width(),
  add_colors = T,
  colors = NightowlOptions$get_colors
)
}
\description{
MISSING_TITLE
}
