% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{frequency}
\alias{frequency}
\title{Calculate Frequency Transformation}
\usage{
frequency(data, mapping)
}
\arguments{
\item{data}{Data frame containing the data to transform}

\item{mapping}{List specifying variable mappings including x variable}
}
\value{
List containing transformed data with frequency column and updated mapping
}
\description{
Transforms count data to frequencies (proportions) within groups
}
