% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bootstrap_mean.R
\name{geom_bootstrap_mean}
\alias{geom_bootstrap_mean}
\title{Bootstrap mean}
\usage{
geom_bootstrap_mean(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  add_ribbon = T,
  add_whiskers = F,
  add_points = F,
  whiskers_width = 0.1,
  ribbon_alpha = 0.5,
  ...
)
}
\description{
Bootstrap mean
}
