% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hillow_median.R
\name{geom_hillow_median}
\alias{geom_hillow_median}
\title{Geom hillow median}
\usage{
geom_hillow_median(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  add_ribbon = T,
  add_whiskers = F,
  add_points = F,
  whiskers_width = 0.1,
  ribbon_alpha = 0.5,
  ...
)
}
\description{
Geom hillow median
}
