% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_section.R
\name{geom_section}
\alias{geom_section}
\title{Title}
\usage{
geom_section(
  mapping = NULL,
  data = NULL,
  ...,
  slope,
  intercept,
  above,
  na.rm = FALSE,
  show.legend = NA
)
}
\arguments{
\item{show.legend}{}
}
\value{

}
\description{
Title
}
