% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bootstrap_mean.R
\docType{data}
\name{GeomBootstrapMean}
\alias{GeomBootstrapMean}
\title{Bootstrap Mean}
\format{
An object of class \code{GeomBootstrapMean} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 8.
}
\usage{
GeomBootstrapMean
}
\description{
Bootstrap Mean
}
\keyword{datasets}
