% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hillow_median.R
\docType{data}
\name{GeomHillowMedian}
\alias{GeomHillowMedian}
\title{Geom Hillow Median}
\format{
An object of class \code{GeomHillowMedian} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 8.
}
\usage{
GeomHillowMedian
}
\description{
Geom Hillow Median
}
\keyword{datasets}
