% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_nightowl_options}
\alias{get_nightowl_options}
\title{Get Nightowl Package Options}
\usage{
get_nightowl_options()
}
\value{
NightowlOptions R6 object instance
}
\description{
Access the global nightowl options instance from package namespace
}
