% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{ggpairs}
\alias{ggpairs}
\title{Boxplot}
\usage{
ggpairs(
  DATA,
  mapping = list(key = NULL, y = NULL, group = NULL, fill = NULL, color = NULL, size =
    NULL, shape = NULL, lty = NULL, id = NULL),
  processing = NULL,
  transform = NULL,
  annotation = NULL,
  facets = NULL,
  label_width = 10,
  svg = NULL,
  diagFn = nightowl::diagFn,
  ...
)
}
\description{
Boxplot
}
