% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{ggplot}
\alias{ggplot}
\title{Create ggplot form list
Setup ggplot
This was difficult, fist store parameters in list,
Convert to symbols
drop the onses which are null, call aes_ function (CAVE: ecex)
also think of other places where params is used, e.g. params$id}
\usage{
ggplot(DATA, aes, ...)
}
\description{
Create ggplot form list
Setup ggplot
This was difficult, fist store parameters in list,
Convert to symbols
drop the onses which are null, call aes_ function (CAVE: ecex)
also think of other places where params is used, e.g. params$id
}
