% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{hide_title}
\alias{hide_title}
\title{Hide plot title (compatible with picasso::hide_title)}
\usage{
hide_title()
}
\value{
ggplot2 theme element
}
\description{
Hide plot title (compatible with picasso::hide_title)
}
