% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{hide_x_axis}
\alias{hide_x_axis}
\title{Hide x-axis from ggplot (compatible with picasso::hide_x_axis)}
\usage{
hide_x_axis()
}
\value{
ggplot2 theme element
}
\description{
Hide x-axis from ggplot (compatible with picasso::hide_x_axis)
}
