% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{hide_y_axis}
\alias{hide_y_axis}
\title{Hide y-axis from ggplot (compatible with picasso::hide_y_axis)}
\usage{
hide_y_axis()
}
\value{
ggplot2 theme element
}
\description{
Hide y-axis from ggplot (compatible with picasso::hide_y_axis)
}
