% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_plots.R
\name{InlinePlot}
\alias{InlinePlot}
\title{R6 Class}
\description{
R6 Class

R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlinePlot-new}{\code{InlinePlot$new()}}
\item \href{#method-InlinePlot-set_html}{\code{InlinePlot$set_html()}}
\item \href{#method-InlinePlot-print}{\code{InlinePlot$print()}}
\item \href{#method-InlinePlot-format}{\code{InlinePlot$format()}}
\item \href{#method-InlinePlot-as.character}{\code{InlinePlot$as.character()}}
\item \href{#method-InlinePlot-get_width}{\code{InlinePlot$get_width()}}
\item \href{#method-InlinePlot-get_height}{\code{InlinePlot$get_height()}}
\item \href{#method-InlinePlot-clone}{\code{InlinePlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlinePlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlinePlot-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlinePlot$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlinePlot-set_html"></a>}}
\if{latex}{\out{\hypertarget{method-InlinePlot-set_html}{}}}
\subsection{Method \code{set_html()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlinePlot$set_html()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlinePlot-print"></a>}}
\if{latex}{\out{\hypertarget{method-InlinePlot-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlinePlot$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlinePlot-format"></a>}}
\if{latex}{\out{\hypertarget{method-InlinePlot-format}{}}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlinePlot$format(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlinePlot-as.character"></a>}}
\if{latex}{\out{\hypertarget{method-InlinePlot-as.character}{}}}
\subsection{Method \code{as.character()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlinePlot$as.character()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlinePlot-get_width"></a>}}
\if{latex}{\out{\hypertarget{method-InlinePlot-get_width}{}}}
\subsection{Method \code{get_width()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlinePlot$get_width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlinePlot-get_height"></a>}}
\if{latex}{\out{\hypertarget{method-InlinePlot-get_height}{}}}
\subsection{Method \code{get_height()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlinePlot$get_height()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlinePlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlinePlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlinePlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
