% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{is_dark}
\alias{is_dark}
\title{Check if color is dark}
\usage{
is_dark(color)
}
\arguments{
\item{color}{Color hex code or name}
}
\value{
Logical indicating if color is dark
}
\description{
Check if color is dark
}
