% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{is_Plot}
\alias{is_Plot}
\title{Is an object a \code{Plot}?}
\usage{
is_Plot(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is an object a \code{Plot}?
}
\examples{
is_Plot(1)
}
