% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{km_summary}
\alias{km_summary}
\title{Generate Kaplan-Meier Summary}
\usage{
km_summary(.formula, data)
}
\arguments{
\item{.formula}{Survival formula object}

\item{data}{Data frame containing the survival data}
}
\value{
HTML formatted summary of survival differences
}
\description{
Creates a formatted summary of survival differences using survdiff
}
