% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{Line plot}
\usage{
line_plot(
  DATA,
  x,
  y,
  fill,
  color,
  id = "USUBJID",
  y_unit = "AVALU",
  title_column = "PARAM",
  add_lines = T,
  lines_alpha = 0.5,
  lines_size = 1,
  add_points = F,
  facet_row = NULL,
  facet_col = NULL,
  method_smooth = NULL,
  color_smooth = "black",
  xlim = NULL,
  ylim = NULL,
  title = NULL,
  ylab = NULL,
  show_se = T,
  log_y = F,
  add_theme = T,
  add_ribbon = F,
  add_whiskers = T,
  theme = ggplot2::theme_light,
  dodge = 0.2,
  points_position = "identity",
  fun.y = NULL,
  scales = "free_y"
)
}
\description{
Line plot
}
