% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{lowerFn}
\alias{lowerFn}
\title{Function for the lower (diagnonal) triangle of the correlation matrix}
\usage{
lowerFn(data, mapping, method = "lm", ...)
}
\description{
Function for the lower (diagnonal) triangle of the correlation matrix
}
