% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_plot.R
\name{make_donut_plot_numeric}
\alias{make_donut_plot_numeric}
\title{MISSING_TITLE}
\usage{
make_donut_plot_numeric(
  data,
  column,
  split,
  colors = NULL,
  brewer_pal = "OrRd",
  color_missing = "grey",
  label = column,
  .range = NULL
)
}
\description{
MISSING_TITLE
}
