% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{model_output}
\alias{model_output}
\title{Summry table}
\usage{
model_output(
  DATA,
  mapping,
  model = "lm",
  options = list(group_split = "row"),
  layers = list(list(type = "tern::summarize_vars", denom = "n")),
  ...
)
}
\description{
Summry table
}
