% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_collapse_top_level}
\alias{nightowl_collapse_top_level}
\title{Collapse top-level list elements (compatible with waRRior::collapse_top_level)}
\usage{
nightowl_collapse_top_level(x)
}
\arguments{
\item{x}{List with nested structure}
}
\value{
Flattened list
}
\description{
Collapse top-level list elements (compatible with waRRior::collapse_top_level)
}
