% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{nightowl_color}
\alias{nightowl_color}
\title{Get a single nightowl color (compatible with picasso::roche_color)}
\usage{
nightowl_color(color = "blue", alpha = 1)
}
\arguments{
\item{color}{Color name (blue, red, green, etc.)}

\item{alpha}{Transparency value (0-1)}
}
\value{
Single color hex code
}
\description{
Get a single nightowl color (compatible with picasso::roche_color)
}
