% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{nightowl_colors}
\alias{nightowl_colors}
\title{Nightowl Color Palette System}
\usage{
nightowl_colors(color = NULL, alpha = 1)
}
\arguments{
\item{color}{Color name (blue, red, green, etc.) or NULL for all colors}

\item{alpha}{Transparency value (0-1)}
}
\value{
Character vector of color hex codes or single color
}
\description{
Internal color palette system to replace external picasso dependency
Based on Roche color scheme from picasso package
}
