% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_drop_columns}
\alias{nightowl_drop_columns}
\title{Drop specified columns from data frame (compatible with waRRior::drop_columns)}
\usage{
nightowl_drop_columns(data, cols)
}
\arguments{
\item{data}{Data frame}

\item{cols}{Column names to drop}
}
\value{
Data frame with specified columns removed
}
\description{
Drop specified columns from data frame (compatible with waRRior::drop_columns)
}
