% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_drop_empty_columns}
\alias{nightowl_drop_empty_columns}
\title{Drop empty columns from data frame (compatible with waRRior::drop_empty_columns)}
\usage{
nightowl_drop_empty_columns(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Data frame with empty columns removed
}
\description{
Drop empty columns from data frame (compatible with waRRior::drop_empty_columns)
}
