% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_fct_lexicographic}
\alias{nightowl_fct_lexicographic}
\title{Factor levels in lexicographic order (compatible with waRRior::fct_lexicographic)}
\usage{
nightowl_fct_lexicographic(x)
}
\arguments{
\item{x}{Factor or character vector}
}
\value{
Factor with levels in lexicographic (alphabetical) order
}
\description{
Factor levels in lexicographic order (compatible with waRRior::fct_lexicographic)
}
