% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_get_groups}
\alias{nightowl_get_groups}
\title{Get groups from grouped data (compatible with waRRior::get_groups)}
\usage{
nightowl_get_groups(data)
}
\arguments{
\item{data}{Grouped data frame}
}
\value{
Vector of group names
}
\description{
Get groups from grouped data (compatible with waRRior::get_groups)
}
