% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_getfun}
\alias{nightowl_getfun}
\title{Get function by name (compatible with waRRior::getfun)}
\usage{
nightowl_getfun(name, package = NULL)
}
\arguments{
\item{name}{Function name as string}

\item{package}{Package name (optional)}
}
\value{
Function object
}
\description{
Get function by name (compatible with waRRior::getfun)
}
