% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_group_split_at}
\alias{nightowl_group_split_at}
\title{Split data at specific groups (compatible with waRRior::named_group_split_at)}
\usage{
nightowl_group_split_at(data, group_var, at)
}
\arguments{
\item{data}{Data frame to split}

\item{group_var}{Grouping variable name}

\item{at}{Group values to split at}
}
\value{
Named list of data frames
}
\description{
Split data at specific groups (compatible with waRRior::named_group_split_at)
}
