% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_length_unique}
\alias{nightowl_length_unique}
\title{Count unique values in a vector (compatible with waRRior::length_unique)}
\usage{
nightowl_length_unique(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Number of unique values
}
\description{
Count unique values in a vector (compatible with waRRior::length_unique)
}
