% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_named_group_split}
\alias{nightowl_named_group_split}
\title{Split data into named groups (compatible with waRRior::named_group_split)}
\usage{
nightowl_named_group_split(data, ...)
}
\arguments{
\item{data}{Data frame to split}

\item{...}{Grouping variables}
}
\value{
Named list of data frames
}
\description{
Split data into named groups (compatible with waRRior::named_group_split)
}
