% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_named_group_split_at}
\alias{nightowl_named_group_split_at}
\title{Split data into named groups at specific values (compatible with waRRior::named_group_split_at)}
\usage{
nightowl_named_group_split_at(
  data,
  group_vars,
  at = NULL,
  keep = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame to split}

\item{group_vars}{Grouping variable names}

\item{at}{Group values to split at (optional)}

\item{keep}{Logical, keep grouping variables in output}

\item{verbose}{Logical, print progress messages}
}
\value{
Named list of data frames
}
\description{
Split data into named groups at specific values (compatible with waRRior::named_group_split_at)
}
