% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{nightowl_palette_discrete}
\alias{nightowl_palette_discrete}
\title{Create discrete color palette for ggplot2}
\usage{
nightowl_palette_discrete(n = NULL)
}
\arguments{
\item{...}{Additional arguments passed to ggplot2::scale_color_manual}
}
\value{
ggplot2 scale object
}
\description{
Create discrete color palette for ggplot2
}
