% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_pop}
\alias{nightowl_pop}
\title{Remove and return last element from list (compatible with waRRior::pop)}
\usage{
nightowl_pop(x)
}
\arguments{
\item{x}{List or vector}
}
\value{
Last element of the list/vector
}
\description{
Remove and return last element from list (compatible with waRRior::pop)
}
