% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{nightowl_reactable_theme}
\alias{nightowl_reactable_theme}
\title{The reactable theme used by nightowl}
\usage{
nightowl_reactable_theme()
}
\value{
A \code{\link[reactable:reactableTheme]{reactable::reactableTheme()}}.
}
\description{
The reactable theme used by nightowl
}
\examples{
render_reactable(head(mtcars), theme = nightowl_reactable_theme())
}
