% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_regex_replace_element_parameter}
\alias{nightowl_regex_replace_element_parameter}
\title{Replace element parameter in string using regex (compatible with waRRior::regex_replace_element_parameter)}
\usage{
nightowl_regex_replace_element_parameter(string, parameter, value)
}
\arguments{
\item{string}{Input string}

\item{parameter}{Parameter name to replace}

\item{value}{New value for the parameter}
}
\value{
Modified string
}
\description{
Replace element parameter in string using regex (compatible with waRRior::regex_replace_element_parameter)
}
