% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_tally_at}
\alias{nightowl_tally_at}
\title{Tally/count values at specified columns (compatible with waRRior::tally_at)}
\usage{
nightowl_tally_at(data, cols)
}
\arguments{
\item{data}{Data frame}

\item{cols}{Column names to tally}
}
\value{
Data frame with counts
}
\description{
Tally/count values at specified columns (compatible with waRRior::tally_at)
}
