% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightowl-plots.R
\name{NightowlPlots}
\alias{NightowlPlots}
\alias{new_NightowlPlots}
\alias{is_NightowlPlots}
\alias{as_ggplot}
\alias{as_html}
\title{A vector of plots}
\usage{
new_NightowlPlots(...)

is_NightowlPlots(x)

as_ggplot(x)

as_html(x)
}
\arguments{
\item{...}{\link{Plot} objects, or a single list of them.}

\item{x}{An object.}
}
\value{
\code{new_NightowlPlots()} returns a \code{NightowlPlots} vector.
\code{is_NightowlPlots()} returns \code{TRUE} or \code{FALSE}. \code{as_ggplot()} returns a
list of ggplot objects. \code{as_html()} returns a list of browsable HTML
blocks.
}
\description{
\code{NightowlPlots} is a \link[vctrs:vctrs]{vctrs} vector whose elements are \link{Plot} objects. It
behaves like any other column: it can live in a tibble, survive
\code{dplyr::mutate()}, \code{filter()} and \code{bind_rows()}, and be concatenated with
\code{c()}. \code{\link[=render_reactable]{render_reactable()}} and \code{\link[=render_kable]{render_kable()}} recognise such columns and
render each element as an inline SVG.
}
\examples{
gg <- ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) + ggplot2::geom_point()
plots <- new_NightowlPlots(Plot$new(gg, svg = list(width = 2, height = 1)))
tibble::tibble(car = "all", trend = plots)
}
