% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R
\name{nw_axis}
\alias{nw_axis}
\title{Apply Axis Transformations}
\usage{
nw_axis(
  g,
  log_x = F,
  log_y = F,
  xlim = NULL,
  ylim = NULL,
  units_x = NULL,
  units_y = NULL,
  ...
)
}
\arguments{
\item{g}{ggplot object to modify axes for}

\item{log_x}{Logical indicating whether to apply log10 transformation to x-axis}

\item{log_y}{Logical indicating whether to apply log10 transformation to y-axis}

\item{xlim}{Numeric vector of x-axis limits}

\item{ylim}{Numeric vector of y-axis limits}

\item{units_x}{Units label for x-axis}

\item{units_y}{Units label for y-axis}

\item{...}{Additional arguments}
}
\value{
Modified ggplot object with axis transformations
}
\description{
Applies axis transformations including log scaling and limits to nightowl plots
}
