% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R
\name{nw_boxplot}
\alias{nw_boxplot}
\title{Add Boxplot Geometry}
\usage{
nw_boxplot(...)
}
\arguments{
\item{...}{Additional arguments passed to add_geom and geom_boxplot}
}
\value{
Modified ggplot object with boxplot layer
}
\description{
Adds boxplot geometry to nightowl plots with position dodging
}
