% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R
\name{nw_colors}
\alias{nw_colors}
\title{Define Color Scales}
\usage{
nw_colors(g, DATA, mapping, ...)
}
\arguments{
\item{g}{ggplot object to apply color scales to}

\item{DATA}{Data frame containing the variables}

\item{mapping}{Aesthetic mapping list containing color/fill variables}

\item{...}{Additional arguments}
}
\value{
Modified ggplot object with color scales applied
}
\description{
Defines color and fill scales for nightowl plots with automatic legend handling
}
