% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R
\name{nw_points}
\alias{nw_points}
\title{Add Points Geometry}
\usage{
nw_points(g, mapping = list(), ...)
}
\arguments{
\item{g}{ggplot object to add points to}

\item{mapping}{Aesthetic mapping list for points}

\item{...}{Additional arguments passed to geom_point}
}
\value{
Modified ggplot object with points layer
}
\description{
Adds point geometry to nightowl plots with custom aesthetic mapping
}
