% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R
\name{nw_smooth}
\alias{nw_smooth}
\title{Add Smooth Trend Line}
\usage{
nw_smooth(g, mapping = list(), dodge = 0, method = "lm", color = "black", ...)
}
\arguments{
\item{g}{ggplot object to add smooth line to}

\item{mapping}{Aesthetic mapping list for smooth line}

\item{dodge}{Position dodge width}

\item{method}{Smoothing method (default: "lm")}

\item{color}{Line color}

\item{...}{Additional arguments passed to geom_smooth}
}
\value{
Modified ggplot object with smooth trend line
}
\description{
Adds smooth trend line to nightowl plots with customizable methods
}
