% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R
\name{nw_summary}
\alias{nw_summary}
\title{Add Summary Statistics Layer}
\usage{
nw_summary(
  g,
  mapping = list(),
  fun.data = NULL,
  fun = NULL,
  dodge = 1,
  shape = 21,
  ...
)
}
\arguments{
\item{g}{ggplot object to add summary to}

\item{mapping}{Aesthetic mapping list for summary statistics}

\item{fun.data}{Function for computing summary data}

\item{fun}{Function for computing single summary statistic}

\item{dodge}{Position dodge width}

\item{shape}{Point shape for summary statistics}

\item{...}{Additional arguments passed to stat_summary}
}
\value{
Modified ggplot object with summary statistics layer
}
\description{
Adds statistical summary layer to nightowl plots with customizable functions
}
