% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{oldboxplot}
\alias{oldboxplot}
\title{Boxplot}
\usage{
oldboxplot(
  DATA,
  x,
  y,
  id,
  group = NULL,
  fill = NULL,
  color = NULL,
  size = NULL,
  shape = NULL,
  facet_col = NULL,
  facet_row = NULL,
  scales = "free_y",
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  label_width = 20,
  axis_text_x_angle = 45,
  axis_text_x_hjust = 1,
  axis_text_x_vjust = 1,
  legend_position = "bottom",
  log_x = F,
  log_y = F,
  theme = picasso::theme_picasso,
  palette_discrete = picasso::roche_palette_discrete(1),
  remove_missing = T,
  plot_height = NULL,
  plot_width = NULL,
  add_boxplot = F,
  add_violin = T,
  add_points = F,
  points_color = picasso::roche_colors("black"),
  points_alpha = 0.5,
  points_position = "identity",
  points_size = 0.5,
  points_stroke = 1,
  dodge = 0.2,
  add_lines = T,
  lines_alpha = 0.5,
  lines_size = 1,
  add_smooth = NULL,
  color_smooth = "black",
  add_ribbon = F,
  add_whiskers = T,
  show_se = T,
  ...
)
}
\description{
Boxplot
}
