% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot}
\alias{Plot}
\title{A rendered plot}
\description{
\code{Plot} wraps a ggplot object together with its SVG rendering options.
Rendering is memoised per object, so \verb{$svg()}, \verb{$html()} and the size
getters share one rendering. \code{Plot} objects are the elements of a
\link{NightowlPlots} vector and therefore what ends up inside table columns.
}
\examples{
p <- Plot$new(
  ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) + ggplot2::geom_point(),
  svg = list(width = 4, height = 3)
)
p$width()
html <- p$html()
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{plot}}{The ggplot object.}

    \item{\code{type}}{A short label used when the object is printed inside a
table column.}

    \item{\code{svg_options}}{List of rendering options merged over the \code{svg}
package option: \code{width}, \code{height}, \code{scaling}, and any argument of
\code{\link[=render_svg]{render_svg()}}.}

    \item{\code{resize}}{Whether the HTML wrapper lets the SVG scale with its
container (\code{TRUE}) or fixes it to its native pixel size.}

    \item{\code{class}}{Extra CSS classes for the HTML wrapper.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Plot-initialize}{\code{Plot$new()}}
    \item \href{#method-Plot-svg}{\code{Plot$svg()}}
    \item \href{#method-Plot-html}{\code{Plot$html()}}
    \item \href{#method-Plot-width}{\code{Plot$width()}}
    \item \href{#method-Plot-height}{\code{Plot$height()}}
    \item \href{#method-Plot-print}{\code{Plot$print()}}
    \item \href{#method-Plot-format}{\code{Plot$format()}}
    \item \href{#method-Plot-as.character}{\code{Plot$as.character()}}
    \item \href{#method-Plot-clone}{\code{Plot$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-initialize}{}}}
\subsection{\code{Plot$new()}}{
  Create a \code{Plot}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Plot$new(plot, svg = list(), type = "Plot", resize = TRUE, class = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{plot}}{A ggplot object.}
      \item{\code{svg}}{List of rendering options, see the \code{svg_options} field.}
      \item{\code{type}}{Label, see the \code{type} field.}
      \item{\code{resize}}{See the \code{resize} field.}
      \item{\code{class}}{See the \code{class} field.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-svg"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-svg}{}}}
\subsection{\code{Plot$svg()}}{
  Render to SVG markup.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Plot$svg(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Rendering options overriding \code{svg_options} for this call.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    An \code{\link[htmltools:HTML]{htmltools::HTML()}} string.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-html"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-html}{}}}
\subsection{\code{Plot$html()}}{
  Render to an HTML block with the nightowl dependency
attached.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Plot$html(resize = self$resize, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{resize}}{Overrides the \code{resize} field.}
      \item{\code{...}}{Passed to \verb{$svg()}.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A browsable \link[htmltools:tag]{htmltools::tag}.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-width"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-width}{}}}
\subsection{\code{Plot$width()}}{
  Width of the rendered SVG in CSS pixels (72 per inch).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Plot$width()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-height"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-height}{}}}
\subsection{\code{Plot$height()}}{
  Height of the rendered SVG in CSS pixels (72 per inch).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Plot$height()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-print"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-print}{}}}
\subsection{\code{Plot$print()}}{
  Print: opens the HTML rendering in the viewer or browser.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Plot$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Ignored.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-format"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-format}{}}}
\subsection{\code{Plot$format()}}{
  One-line description used inside tables.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Plot$format(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Ignored.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-as.character"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-as.character}{}}}
\subsection{\code{Plot$as.character()}}{
  The SVG markup as a plain string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Plot$as.character(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Passed to \verb{$svg()}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-clone}{}}}
\subsection{\code{Plot$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Plot$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
