% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{plot_grouped_km}
\alias{plot_grouped_km}
\title{MISSING_TITLE}
\usage{
plot_grouped_km(
  data,
  time,
  event,
  treatment,
  covariates = NULL,
  split = NULL,
  width = "75vw",
  flex_direction = "column",
  as_ggplot = FALSE,
  title = function() {
     glue::glue("{.split}", .envir = parent.frame())
 },
  subtitle = function() NULL,
  note = function() NULL,
  ...
)
}
\description{
MISSING_TITLE
}
