% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{plot_grouped_km_compact}
\alias{plot_grouped_km_compact}
\title{Plot Compact Grouped Kaplan-Meier Curves}
\usage{
plot_grouped_km_compact(
  data,
  time,
  event,
  treatment,
  covariates = NULL,
  split,
  width = "400px",
  add_table = TRUE,
  break_width = (max(data[[time]], na.rm = T)/6),
  add_p = TRUE,
  ...
)
}
\arguments{
\item{data}{Data frame containing the survival data}
}
\value{
HTML object with compact arranged Kaplan-Meier plots
}
\description{
Creates compact multiple Kaplan-Meier curves with minimal styling
}
