% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{plot_km_covariates}
\alias{plot_km_covariates}
\title{Plot Kaplan-Meier with Covariate Distribution}
\usage{
plot_km_covariates(
  data,
  time,
  event,
  treatment,
  covariates = NULL,
  title = event,
  subtitle = NULL,
  landmark = NULL,
  ylab = "Survival",
  xlab = "Time (Days)",
  add_p = TRUE,
  add_table = TRUE,
  add_summary = FALSE,
  add_median = TRUE,
  wrap = NULL,
  legend_position = "top",
  height = 600,
  width = "100\%",
  colors = unname(unlist(nightowl::nightowl_colors())),
  lowrider_theme = "print",
  as_ggplot = FALSE,
  note = NULL,
  break_width = (max(data[[time]], na.rm = T)/20)
)
}
\arguments{
\item{data}{Data frame containing the survival data}
}
\value{
Interactive plot combining survival curves and covariate distributions
}
\description{
Creates Kaplan-Meier curves with additional panels showing covariate distributions over time
}
