% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_percentages.R
\name{plot_stacked_percentages}
\alias{plot_stacked_percentages}
\title{Plot stacked percentage barplot}
\usage{
plot_stacked_percentages(
  DATA,
  x,
  y,
  fill = y,
  title = "Missing title",
  fill_colors = NULL,
  debug = F,
  position = "stack",
  facet_col = NULL,
  facet_row = NULL,
  facet_wrap = NULL,
  add_labels = T,
  add_theme = T,
  .theme = ggplot2::theme_minimal,
  legend_title = NULL,
  dodge = 0,
  x_angle = NULL,
  strip_x_angle = NULL,
  strip_y_angle = NULL,
  reverse = F,
  data_id_prefix = stringi::stri_reverse(AceOfSpades::random_id()),
  onclick = "console.log($(\\"[class^=\\"hover_svg\\"]\\"))",
  explicit_na = T
)
}
\description{
Plot stacked percentage barplot
}
