% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{prepare_data_for_plotting}
\alias{prepare_data_for_plotting}
\title{Handle missing values}
\usage{
prepare_data_for_plotting(
  DATA,
  cols = NULL,
  remove_missing = T,
  to_factor = TRUE
)
}
\description{
Handle missing values
}
