% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{print.nightowl_error}
\alias{print.nightowl_error}
\title{Print method for nightowl errors}
\usage{
\method{print}{nightowl_error}(x, ...)
}
\arguments{
\item{x}{A nightowl error object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for nightowl errors
}
