% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{print.security_error}
\alias{print.security_error}
\title{Print method for security errors}
\usage{
\method{print}{security_error}(x, ...)
}
\arguments{
\item{x}{A security error object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for security errors
}
