% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{print.validation_error}
\alias{print.validation_error}
\title{Print method for validation errors}
\usage{
\method{print}{validation_error}(x, ...)
}
\arguments{
\item{x}{A validation error object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for validation errors
}
