% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{reactable_summary}
\alias{reactable_summary}
\title{MISSING_TITLE}
\usage{
reactable_summary(
  data,
  split = NULL,
  columns,
  group_by = NULL,
  plan = "sequential",
  ...
)
}
\value{

}
\description{
MISSING_TITLE
}
