% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.r
\name{render_girafe}
\alias{render_girafe}
\title{MISSING_TITLE}
\usage{
render_girafe(
  code,
  ggobj = NULL,
  pointsize = 12,
  width_svg = NULL,
  height_svg = NULL,
  options = list(),
  fix_rect = TRUE,
  ...
)
}
\description{
MISSING_TITLE
}
