% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_nightowl_options}
\alias{set_nightowl_options}
\title{Set Nightowl Package Options}
\usage{
set_nightowl_options(options)
}
\arguments{
\item{options}{NightowlOptions R6 object instance}
}
\description{
Replace the global nightowl options instance
}
