% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bootstrap_mean.R
\docType{data}
\name{StatBootstrapMean}
\alias{StatBootstrapMean}
\title{Stat Bootstrap mean}
\format{
An object of class \code{StatBootstrapMean} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatBootstrapMean
}
\description{
Stat Bootstrap mean
}
\keyword{datasets}
