% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hillow_median.R
\docType{data}
\name{StatHillowMedian}
\alias{StatHillowMedian}
\title{Stat Hillow Median}
\format{
An object of class \code{StatHillowMedian} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatHillowMedian
}
\description{
Stat Hillow Median
}
\keyword{datasets}
