% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table}
\alias{table}
\title{Summry table}
\usage{
table(
  DATA,
  mapping,
  options = list(group_split = "row"),
  layers = list(list(type = "tern::summarize_vars", denom = "n")),
  ...
)
}
\description{
Summry table
}
