% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{Test}
\alias{Test}
\title{R6 Class}
\description{
R6 Class

R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Test-new}{\code{Test$new()}}
\item \href{#method-Test-clone}{\code{Test$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Test-new"></a>}}
\if{latex}{\out{\hypertarget{method-Test-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Test$new(.data, y, x = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Test-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Test-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Test$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
