% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{throw_data_error}
\alias{throw_data_error}
\title{Throw a data error}
\usage{
throw_data_error(message, data_info = NULL, call = sys.call(-1))
}
\arguments{
\item{message}{Error message}

\item{data_info}{Information about the problematic data}

\item{call}{The call that generated the error (defaults to parent call)}
}
\description{
Throw a data error
}
\keyword{internal}
