% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{throw_security_error}
\alias{throw_security_error}
\title{Throw a security error}
\usage{
throw_security_error(message, operation = NULL, call = sys.call(-1))
}
\arguments{
\item{message}{Error message}

\item{operation}{The operation that was attempted}

\item{call}{The call that generated the error (defaults to parent call)}
}
\description{
Throw a security error
}
\keyword{internal}
