% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{upperFn}
\alias{upperFn}
\title{Function for the upper (diagnonal) triangle of the correlation matrix}
\usage{
upperFn(data, mapping, method = "lm", ...)
}
\description{
Function for the upper (diagnonal) triangle of the correlation matrix
}
