% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_plots.R
\name{add_forestplot}
\alias{add_forestplot}
\title{MISSING_TITLE}
\usage{
add_forestplot(
  x,
  fun_data = ggplot2::mean_cl_boot,
  xintercept = NULL,
  xlim = NULL
)
}
\value{

}
\description{
MISSING_TITLE
}
