% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_plots.R
\name{add_inline_pointrange}
\alias{add_inline_pointrange}
\title{MISSING_TITLE}
\usage{
add_inline_pointrange(
  x,
  fun_data = NULL,
  mapping = list(x = "y", xmin = "ymin", xmax = "ymax", y = NULL),
  ...
)
}
\description{
MISSING_TITLE
}
