% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_plots.R
\name{add_violin}
\alias{add_violin}
\title{MISSING_TITLE}
\usage{
add_violin(
  x,
  ylim = NULL,
  height = 0.3,
  width = 3.5,
  scaling = 1,
  theme = ggplot2::theme_void,
  hide_legend = T,
  hide_x_axis = T,
  hide_y_axis = T,
  fill = "#B9B9B8",
  expansion_y = 10,
  add_download_button = FALSE,
  fun.data = ggplot2::mean_cl_boot
)
}
\description{
MISSING_TITLE
}
