% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_plots.R
\name{add_violin}
\alias{add_violin}
\title{MISSING_TITLE}
\usage{
add_violin(
  x,
  ylim = NULL,
  height = 0.8,
  width = 8,
  scaling = 1,
  theme = ggplot2::theme_void,
  hide_legend = T,
  hide_x_axis = T,
  hide_y_axis = T,
  fill = picasso::roche_colors("lightblue"),
  expansion_y = 0,
  add_download_button = FALSE,
  fun.data = ggplot2::mean_cl_boot
)
}
\description{
MISSING_TITLE
}
