% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Boxplot}
\usage{
boxplot(
  DATA,
  x,
  y,
  fill = NULL,
  color = NULL,
  title = NULL,
  xlab = NULL,
  axis.text.x.angle = 45,
  axis.text.x.hjust = 1,
  axis.text.x.vjust = 1,
  legend.position = "bottom",
  x_label_width = 20,
  ylab = NULL,
  log_y = F,
  remove_missing = T,
  dodge = 1,
  add_violin = T,
  add_boxplot = T,
  add_points = T,
  points_color = picasso::roche_color("black"),
  points_size = 1,
  points_stroke = 1,
  points_alpha = 0.8,
  add_title = T,
  facet_row = NULL,
  facet_col = NULL,
  label_width = 20,
  theme = picasso::theme_picasso,
  palette_discrete = picasso::roche_palette_discrete(1),
  scales = "free_x",
  ...
)
}
\description{
Boxplot
}
