% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{calc_summary}
\alias{calc_summary}
\title{MISSING_TITLE}
\usage{
calc_summary(
  data,
  column,
  calculations = list(N. = nightowl::n),
  unnest = TRUE,
  names_sep = "."
)
}
\value{

}
\description{
MISSING_TITLE
}
