% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{calc_summary_numeric}
\alias{calc_summary_numeric}
\title{MISSING_TITLE}
\usage{
calc_summary_numeric(
  data,
  column,
  calculations = list(N. = length, Missing = function(x) {     return(sum(is.na(x))) },
    Median = function(x) median(x, na.rm = T), Mean = nightowl::formated_mean, Forestplot
    = nightowl::add_forestplot),
  parameters = list(Forestplot = list(xintercept = mean(data[[column]], na.rm = T),
    xlim = c(mean(data[[column]], na.rm = T) - IQR(data[[column]], na.rm = T),
    mean(data[[column]], na.rm = T) + IQR(data[[column]], na.rm = T)))),
  unnest = TRUE
)
}
\value{

}
\description{
MISSING_TITLE
}
