% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{create_Surv_formula}
\alias{create_Surv_formula}
\title{MISSING_TITLE}
\usage{
create_Surv_formula(
  data,
  time,
  event,
  treatment,
  covariates = NULL,
  strata = NULL,
  random_effects = NULL
)
}
\value{

}
\description{
MISSING_TITLE
}
