% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{DeclarativePlot}
\alias{DeclarativePlot}
\title{R6 Class}
\description{
R6 Class

R6 Class
}
\section{Super class}{
\code{\link[nightowl:Plot]{nightowl::Plot}} -> \code{DeclarativePlot}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeclarativePlot-new}{\code{DeclarativePlot$new()}}
\item \href{#method-DeclarativePlot-select_data}{\code{DeclarativePlot$select_data()}}
\item \href{#method-DeclarativePlot-transform_data}{\code{DeclarativePlot$transform_data()}}
\item \href{#method-DeclarativePlot-prepare_facets}{\code{DeclarativePlot$prepare_facets()}}
\item \href{#method-DeclarativePlot-set_plot}{\code{DeclarativePlot$set_plot()}}
\item \href{#method-DeclarativePlot-format}{\code{DeclarativePlot$format()}}
\item \href{#method-DeclarativePlot-clone}{\code{DeclarativePlot$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="nightowl" data-topic="Plot" data-id="as.character"><a href='../../nightowl/html/Plot.html#method-Plot-as.character'><code>nightowl::Plot$as.character()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nightowl" data-topic="Plot" data-id="get_height"><a href='../../nightowl/html/Plot.html#method-Plot-get_height'><code>nightowl::Plot$get_height()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nightowl" data-topic="Plot" data-id="get_width"><a href='../../nightowl/html/Plot.html#method-Plot-get_width'><code>nightowl::Plot$get_width()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nightowl" data-topic="Plot" data-id="html"><a href='../../nightowl/html/Plot.html#method-Plot-html'><code>nightowl::Plot$html()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nightowl" data-topic="Plot" data-id="parse_style"><a href='../../nightowl/html/Plot.html#method-Plot-parse_style'><code>nightowl::Plot$parse_style()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nightowl" data-topic="Plot" data-id="print"><a href='../../nightowl/html/Plot.html#method-Plot-print'><code>nightowl::Plot$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="nightowl" data-topic="Plot" data-id="svg"><a href='../../nightowl/html/Plot.html#method-Plot-svg'><code>nightowl::Plot$svg()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclarativePlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-DeclarativePlot-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclarativePlot$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclarativePlot-select_data"></a>}}
\if{latex}{\out{\hypertarget{method-DeclarativePlot-select_data}{}}}
\subsection{Method \code{select_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclarativePlot$select_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclarativePlot-transform_data"></a>}}
\if{latex}{\out{\hypertarget{method-DeclarativePlot-transform_data}{}}}
\subsection{Method \code{transform_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclarativePlot$transform_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclarativePlot-prepare_facets"></a>}}
\if{latex}{\out{\hypertarget{method-DeclarativePlot-prepare_facets}{}}}
\subsection{Method \code{prepare_facets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclarativePlot$prepare_facets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclarativePlot-set_plot"></a>}}
\if{latex}{\out{\hypertarget{method-DeclarativePlot-set_plot}{}}}
\subsection{Method \code{set_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclarativePlot$set_plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclarativePlot-format"></a>}}
\if{latex}{\out{\hypertarget{method-DeclarativePlot-format}{}}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclarativePlot$format(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeclarativePlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeclarativePlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeclarativePlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
