% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{forestplot}
\alias{forestplot}
\title{MISSING_TITLE}
\usage{
forestplot(
  x,
  xmin,
  xmax,
  xlim = NULL,
  xintercept = NULL,
  height = 0.5,
  xlab = NULL,
  ylab = NULL,
  hide_y_axis = TRUE,
  hide_x_axis = TRUE,
  hide_legend = TRUE,
  theme = ggplot2::theme_void
)
}
\description{
MISSING_TITLE
}
