% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{format_frequencies}
\alias{format_frequencies}
\title{MISSING_TITLE}
\usage{
format_frequencies(
  x,
  N = length(x),
  output = "print",
  digits = 1,
  str_width = NightowlOptions$get_header_width(),
  add_legend = FALSE,
  add_colors = T,
  colors = NightowlOptions$get_colors
)
}
\description{
MISSING_TITLE
}
