% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{ggpairs}
\alias{ggpairs}
\title{Boxplot}
\usage{
ggpairs(
  DATA,
  key,
  value = NULL,
  id = NULL,
  group = NULL,
  fill = NULL,
  color = NULL,
  size = NULL,
  shape = NULL,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  label_width = 20,
  axis_text_x_angle = 45,
  axis_text_x_hjust = 1,
  axis_text_x_vjust = 1,
  legend_position = "bottom",
  remove_missing = T,
  ...
)
}
\description{
Boxplot
}
