% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{is_Plot}
\alias{is_Plot}
\title{Check if object is a Plot instance}
\usage{
is_Plot(x)
}
\arguments{
\item{x}{Any R object to test}
}
\value{
Logical. TRUE if object is a Plot instance, FALSE otherwise
}
\description{
Tests whether an object inherits from both the Plot and R6 classes,
confirming it is a valid Plot object.
}
