% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{Line plot}
\usage{
line_plot(
  DATA,
  x,
  y,
  fill = NULL,
  color = NULL,
  id = "USUBJID",
  add_lines = T,
  lines_alpha = 0.5,
  lines_size = 1,
  add_points = F,
  facet_row = NULL,
  facet_col = NULL,
  facet_label_width = 20,
  method_smooth = NULL,
  color_smooth = "black",
  xlim = NULL,
  ylim = NULL,
  title = NULL,
  ylab = NULL,
  axis.text.x.angle = 45,
  show_se = T,
  log_y = F,
  add_ribbon = F,
  add_whiskers = T,
  theme = picasso::theme_picasso,
  palette_discrete = picasso::roche_palette_discrete(1),
  dodge = 0.2,
  points_position = "identity",
  summarise_y = NULL,
  scales = "free_y",
  plot_height = NULL,
  plot_width = NULL,
  ...
)
}
\description{
Line plot
}
