% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_plot.R
\name{make_donut_plot_categorical}
\alias{make_donut_plot_categorical}
\title{MISSING_TITLE}
\usage{
make_donut_plot_categorical(
  data,
  column,
  split,
  threshold_levels = 3,
  threshold_fractions = 0.089,
  colors = NULL,
  brewer_pal = "OrRd",
  color_missing = "#BEBEBE",
  label = column
)
}
\description{
MISSING_TITLE
}
