% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nightowl_pop}
\alias{nightowl_pop}
\title{Remove specified elements from vector (compatible with waRRior::pop)}
\usage{
nightowl_pop(x, remove)
}
\arguments{
\item{x}{Vector or list}

\item{remove}{Elements to remove from x}
}
\value{
Vector x with elements in remove excluded
}
\description{
Remove specified elements from vector (compatible with waRRior::pop)
}
