% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot}
\alias{Plot}
\title{R6 Class}
\description{
R6 Class

R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Plot-new}{\code{Plot$new()}}
\item \href{#method-Plot-select_data}{\code{Plot$select_data()}}
\item \href{#method-Plot-transform_data}{\code{Plot$transform_data()}}
\item \href{#method-Plot-prepare_facets}{\code{Plot$prepare_facets()}}
\item \href{#method-Plot-print}{\code{Plot$print()}}
\item \href{#method-Plot-clone}{\code{Plot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-new"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-select_data"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-select_data}{}}}
\subsection{Method \code{select_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$select_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-transform_data"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-transform_data}{}}}
\subsection{Method \code{transform_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$transform_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-prepare_facets"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-prepare_facets}{}}}
\subsection{Method \code{prepare_facets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$prepare_facets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-print"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
