% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot}
\alias{Plot}
\title{R6 Class}
\description{
R6 Class

R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Plot-new}{\code{Plot$new()}}
\item \href{#method-Plot-select_data}{\code{Plot$select_data()}}
\item \href{#method-Plot-transform_data}{\code{Plot$transform_data()}}
\item \href{#method-Plot-prepare_facets}{\code{Plot$prepare_facets()}}
\item \href{#method-Plot-set_ggplot}{\code{Plot$set_ggplot()}}
\item \href{#method-Plot-set_options_svg}{\code{Plot$set_options_svg()}}
\item \href{#method-Plot-html}{\code{Plot$html()}}
\item \href{#method-Plot-set_render_svg}{\code{Plot$set_render_svg()}}
\item \href{#method-Plot-print}{\code{Plot$print()}}
\item \href{#method-Plot-format}{\code{Plot$format()}}
\item \href{#method-Plot-as.character}{\code{Plot$as.character()}}
\item \href{#method-Plot-get_width}{\code{Plot$get_width()}}
\item \href{#method-Plot-get_height}{\code{Plot$get_height()}}
\item \href{#method-Plot-inspect}{\code{Plot$inspect()}}
\item \href{#method-Plot-clone}{\code{Plot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-new"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-select_data"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-select_data}{}}}
\subsection{Method \code{select_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$select_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-transform_data"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-transform_data}{}}}
\subsection{Method \code{transform_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$transform_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-prepare_facets"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-prepare_facets}{}}}
\subsection{Method \code{prepare_facets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$prepare_facets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-set_ggplot"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-set_ggplot}{}}}
\subsection{Method \code{set_ggplot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$set_ggplot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-set_options_svg"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-set_options_svg}{}}}
\subsection{Method \code{set_options_svg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$set_options_svg(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-html"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-html}{}}}
\subsection{Method \code{html()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$html(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-set_render_svg"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-set_render_svg}{}}}
\subsection{Method \code{set_render_svg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$set_render_svg()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-print"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-format"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-format}{}}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$format(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-as.character"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-as.character}{}}}
\subsection{Method \code{as.character()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$as.character()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-get_width"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-get_width}{}}}
\subsection{Method \code{get_width()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$get_width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-get_height"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-get_height}{}}}
\subsection{Method \code{get_height()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$get_height()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-inspect"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-inspect}{}}}
\subsection{Method \code{inspect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$inspect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
