% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Boxplot}
\usage{
plot(
  DATA,
  mapping = list(x = NULL, y = NULL, group = NULL, fill = NULL, color = NULL, size =
    NULL, shape = NULL, lty = NULL, id = NULL),
  processing = NULL,
  transform = NULL,
  layers = list(),
  annotation = NULL,
  axis = NULL,
  theming = NULL,
  facets = NULL,
  svg = NULL,
  dodge = 1,
  ...
)
}
\description{
Boxplot
}
