% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot}
\alias{Plot}
\title{R6 Class}
\description{
R6 Class

R6 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ggplot}}{object, potentially other plot types possible (base does not work)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ggplot}}{object, potentially other plot types possible (base does not work)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Plot-new}{\code{Plot$new()}}
\item \href{#method-Plot-svg}{\code{Plot$svg()}}
\item \href{#method-Plot-parse_style}{\code{Plot$parse_style()}}
\item \href{#method-Plot-html}{\code{Plot$html()}}
\item \href{#method-Plot-print}{\code{Plot$print()}}
\item \href{#method-Plot-format}{\code{Plot$format()}}
\item \href{#method-Plot-as.character}{\code{Plot$as.character()}}
\item \href{#method-Plot-get_width}{\code{Plot$get_width()}}
\item \href{#method-Plot-get_height}{\code{Plot$get_height()}}
\item \href{#method-Plot-clone}{\code{Plot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-new"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$new(
  plot = NULL,
  options_svg = NULL,
  type = "NightowlPlot",
  resize = TRUE,
  css = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-svg"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-svg}{}}}
\subsection{Method \code{svg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$svg(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-parse_style"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-parse_style}{}}}
\subsection{Method \code{parse_style()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$parse_style(style = self$css$style)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-html"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-html}{}}}
\subsection{Method \code{html()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$html(resize = self$resize)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-print"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$print(browser = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-format"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-format}{}}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$format(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-as.character"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-as.character}{}}}
\subsection{Method \code{as.character()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$as.character()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-get_width"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-get_width}{}}}
\subsection{Method \code{get_width()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$get_width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-get_height"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-get_height}{}}}
\subsection{Method \code{get_height()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$get_height()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Plot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
