% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Boxplot}
\usage{
plot(
  DATA,
  processing = list(remove_missing = T),
  mapping = list(x = NULL, y = NULL, group = NULL, fill = NULL, color = NULL, size =
    NULL, shape = NULL, lty = NULL, id = NULL),
  facetting = NULL,
  annotation = NULL,
  axis = NULL,
  theming = NULL,
  points = NULL,
  dotplot = NULL,
  lines = NULL,
  boxplot = NULL,
  violin = NULL,
  smooth = NULL,
  summary = NULL,
  traces = NULL,
  facets = NULL,
  dodge = NULL,
  svg = NULL,
  ...
)
}
\description{
Boxplot
}
