% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{plot_coxph}
\alias{plot_coxph}
\title{MISSING_TITLE}
\usage{
plot_coxph(
  data,
  time,
  event,
  treatment,
  covariates = NULL,
  strata = NULL,
  engine = "kable",
  kable_style = kableExtra::kable_paper,
  split = NULL,
  title = NULL,
  ...
)
}
\description{
MISSING_TITLE
}
