% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{plot_coxph}
\alias{plot_coxph}
\title{MISSING_TITLE}
\usage{
plot_coxph(
  data,
  time,
  event,
  treatment,
  covariates = NULL,
  show_only_treatment = FALSE,
  strata = NULL,
  engine = "kable",
  kable_style = kableExtra::kable_paper,
  split = NULL,
  conf_range = NULL,
  label_left = "Comparison better",
  label_right = "Reference better",
  title = "",
  ...
)
}
\description{
MISSING_TITLE
}
