% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_chisq.R
\name{plot_grouped_chisq}
\alias{plot_grouped_chisq}
\title{MISSING_TITLE}
\usage{
plot_grouped_chisq(
  df,
  split_by,
  x,
  y,
  test = NULL,
  order_by = "p.value",
  pal = NULL,
  width = 3,
  height = 3,
  flex_direction = "row",
  flex_wrap = "wrap",
  ...
)
}
\description{
MISSING_TITLE
}
