% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{plot_grouped_km_compact}
\alias{plot_grouped_km_compact}
\title{MISSING_TITLE}
\usage{
plot_grouped_km_compact(
  data,
  time,
  event,
  treatment,
  covariates = NULL,
  split,
  width = "300px",
  add_table = FALSE,
  add_p = TRUE,
  ...
)
}
\value{

}
\description{
MISSING_TITLE
}
