% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{plot_km}
\alias{plot_km}
\title{MISSING_TITLE}
\usage{
plot_km(
  data,
  time,
  event,
  treatment,
  covariates = NULL,
  title = event,
  subtitle = NULL,
  landmark = NULL,
  ylab = "Survival",
  xlab = "Time (Days)",
  add_p = TRUE,
  add_table = TRUE,
  add_summary = FALSE,
  add_median = TRUE,
  wrap = NULL,
  legend_position = "top",
  height = 600,
  table_height_fraction = 0.2,
  width = "100\%",
  colors = unname(unlist(picasso::roche_colors())),
  lowrider_theme = "print",
  as_ggplot = FALSE,
  note = NULL,
  break_width = (max(data[[time]], na.rm = T)/20)
)
}
\description{
MISSING_TITLE
}
