% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_chisq.R
\name{reactable_grouped_chisq}
\alias{reactable_grouped_chisq}
\title{MISSING_TITLE}
\usage{
reactable_grouped_chisq(
  df,
  split_by,
  x,
  y,
  test = NULL,
  order_by = "p.value",
  pal = NULL,
  width = 3,
  height = 3,
  ...
)
}
\description{
MISSING_TITLE
}
