% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{render_kable}
\alias{render_kable}
\title{MISSING_TITLE}
\usage{
render_kable(
  .tbl,
  caption = NULL,
  full_width = FALSE,
  digits = 2,
  add_scale = T,
  column_width = "20em",
  ...
)
}
\value{

}
\description{
MISSING_TITLE
}
