% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{render_reactable}
\alias{render_reactable}
\title{MISSING_TITLE}
\usage{
render_reactable(
  .tbl,
  html_columns = NULL,
  theme = function() reactablefmtr::nytimes(centered = TRUE),
  digits = 2,
  defaultPageSize = 10,
  filterable = T,
  minWidth_default = 100,
  maxWidth_default = NULL,
  minWidth_html = 300,
  maxWidth_html = NULL,
  fullWidth = FALSE,
  ...
)
}
\description{
MISSING_TITLE
}
