% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{render_reactable}
\alias{render_reactable}
\title{MISSING_TITLE}
\usage{
render_reactable(
  tibble,
  html_columns = NULL,
  theme = "table",
  defaultPageSize = 10,
  minWidth_html = 300,
  fullWidth = FALSE,
  ...
)
}
\value{

}
\description{
MISSING_TITLE
}
