% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{render_reactable}
\alias{render_reactable}
\title{MISSING_TITLE}
\usage{
render_reactable(
  .tbl,
  html_columns = NULL,
  theme = function() reactablefmtr::nytimes(centered = TRUE),
  digits = 2,
  defaultPageSize = 10,
  filterable = T,
  add_scale = T,
  defaultColDef = list(header = function(value) gsub(".", " ", value, fixed = TRUE),
    align = "center", na = "-", html = TRUE),
  fullWidth = FALSE,
  ...
)
}
\description{
MISSING_TITLE
}
