% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.r
\name{render_svg}
\alias{render_svg}
\title{Create SVG string}
\usage{
render_svg(
  g,
  height = 8,
  width = 8,
  element_width = "100\%",
  element_height = "100\%",
  scaling = 1,
  add_download_button = T,
  standalone = F,
  bg = "transparent",
  font_family = "Lato, sans-serif",
  fix_rect = T,
  ...
)
}
\arguments{
\item{g}{ggplot object}
}
\description{
Create SVG string
}
