% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarise}
\alias{summarise}
\title{MISSING_TITLE}
\usage{
summarise(
  data,
  column,
  template = NULL,
  calculations = template$calculations,
  parameters = template$parameters,
  unnest = TRUE,
  name_for_column = "Variable",
  names_sep = "."
)
}
\description{
MISSING_TITLE
}
