% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarise}
\alias{summarise}
\title{MISSING_TITLE}
\usage{
summarise(
  data,
  column,
  calculations = list(N. = length),
  parameters = list(),
  unnest = TRUE,
  name_for_column = "Variable",
  names_sep = "."
)
}
\value{

}
\description{
MISSING_TITLE
}
