% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarise_categorical}
\alias{summarise_categorical}
\title{MISSING_TITLE}
\usage{
summarise_categorical(
  data,
  column,
  calculations = list(N. = length, Freq = nightowl::format_frequencies),
  parameters = list(Freq = list(add_colors = FALSE)),
  unnest = TRUE,
  names_sep = NULL
)
}
\description{
MISSING_TITLE
}
