% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarise_categorical_barplot}
\alias{summarise_categorical_barplot}
\title{MISSING_TITLE}
\usage{
summarise_categorical_barplot(
  data,
  column,
  calculations = list(N. = length, Freq = nightowl::format_frequencies, Barplot =
    nightowl::add_barplot),
  parameters = list(Freq = list(add_colors = TRUE)),
  unnest = TRUE,
  names_sep = NULL
)
}
\description{
MISSING_TITLE
}
