% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarise_numeric}
\alias{summarise_numeric}
\title{MISSING_TITLE}
\usage{
summarise_numeric(
  data,
  column,
  calculations = list(N. = length, Missing = function(x) {
     return(sum(is.na(x)))
 },
    `Extreme Values` = nightowl::count_outliers, Median = function(x) median(x, na.rm =
    T), Min = function(x) min(x, na.rm = T), Max = function(x) max(x, na.rm = T), Mean =
    nightowl::formated_mean),
  parameters = list(),
  unnest = TRUE
)
}
\description{
MISSING_TITLE
}
