% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarise_numeric_histogram}
\alias{summarise_numeric_histogram}
\title{MISSING_TITLE}
\usage{
summarise_numeric_histogram(
  data,
  column,
  calculations = list(N. = length, Median = function(x) median(x, na.rm = T), Mean =
    nightowl::formated_mean, Histogram = nightowl::add_histogram),
  parameters = list(),
  unnest = TRUE
)
}
\description{
MISSING_TITLE
}
