% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarise_numeric_pointrange}
\alias{summarise_numeric_pointrange}
\title{MISSING_TITLE}
\usage{
summarise_numeric_pointrange(
  data,
  column,
  calculations = list(N. = length, Median = function(x) median(x, na.rm = T), Mean =
    nightowl::formated_mean, Pointrange = nightowl::add_inline_pointrange),
  parameters = list(Pointrange = list(fun_data = ggplot2::mean_cl_boot, xintercept =
    mean(data[[column]], na.rm = T), xlim = c(min(data[[column]], na.rm = T),
    max(data[[column]], na.rm = T)))),
  unnest = TRUE
)
}
\description{
MISSING_TITLE
}
