% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summarise_numeric_violin}
\alias{summarise_numeric_violin}
\title{MISSING_TITLE}
\usage{
summarise_numeric_violin(
  self,
  calculations = list(N. = length, Median = function(x) median(x, na.rm = T), Mean =
    nightowl::formated_mean, Violin = nightowl::add_violin),
  parameters = list(Violin = list(theme = picasso::theme_void, ylim =
    range(self$data[[self$column]], na.rm = T)))
)
}
\description{
MISSING_TITLE
}
