% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.r
\name{summary}
\alias{summary}
\title{Summary (Alias)}
\usage{
summary(
  .data,
  column,
  group_by = NULL,
  method = NULL,
  labels = NULL,
  debug = F,
  ...
)
}
\arguments{
\item{...}{All arguments passed to new_summary}
}
\value{
A Summary R6 object
}
\description{
Alias for new_summary function to maintain backward compatibility
}
