% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.r, R/add_plots.R
\name{summary}
\alias{summary}
\title{Summarise}
\usage{
summary(
  g,
  mapping = list(),
  fun.data = NULL,
  fun = NULL,
  dodge = 1,
  shape = 21,
  ...
)

summary(
  g,
  mapping = list(),
  fun.data = NULL,
  fun = NULL,
  dodge = 1,
  shape = 21,
  ...
)
}
\description{
Summarise

Add summary
}
