% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R, R/summaries.R
\name{summary}
\alias{summary}
\title{Add summary}
\usage{
summary(
  data,
  column,
  group_by = NULL,
  .mean = ggplot2::mean_cl_boot,
  add_caption = TRUE,
  add_forest = TRUE,
  wrap_header = TRUE,
  output = "raw",
  ...
)

summary(
  data,
  column,
  group_by = NULL,
  .mean = ggplot2::mean_cl_boot,
  add_caption = TRUE,
  add_forest = TRUE,
  wrap_header = TRUE,
  output = "raw",
  ...
)
}
\value{

}
\description{
Add summary

MISSING_TITLE
}
