% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R, R/summaries.R
\name{summary}
\alias{summary}
\title{Add summary}
\usage{
summary(
  data,
  column,
  group_by = NULL,
  .mean = ggplot2::mean_cl_boot,
  .summarise_numeric = nightowl::summarise_numeric_forestplot,
  .summarise_categorical = nightowl::summarise_categorical_barplot,
  calc_p = TRUE,
  show_p = TRUE,
  add_caption = TRUE,
  wrap_header = TRUE,
  output = "raw",
  labels = NULL,
  .range = NULL,
  ...
)

summary(
  data,
  column,
  group_by = NULL,
  .mean = ggplot2::mean_cl_boot,
  .summarise_numeric = nightowl::summarise_numeric_forestplot,
  .summarise_categorical = nightowl::summarise_categorical_barplot,
  calc_p = TRUE,
  show_p = TRUE,
  add_caption = TRUE,
  wrap_header = TRUE,
  output = "raw",
  labels = NULL,
  .range = NULL,
  ...
)
}
\value{

}
\description{
Add summary

MISSING_TITLE
}
