% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{summary_table}
\alias{summary_table}
\title{Summry table
Table to display data that is being used in plots. All character columns
are converted into factors.}
\usage{
summary_table(
  DATA,
  x,
  y,
  group = NULL,
  group_split = "rows",
  facet_cols = NULL,
  facet_rows = NULL,
  remove_missing = FALSE,
  label_width = 20,
  denom = "n",
  ...
)
}
\arguments{
\item{x}{Split columns by}

\item{y}{Variable to summarise}

\item{group}{Grouping variable}

\item{group_split}{Split Groups be split into "cols" or "rows"}

\item{facet_cols}{Split columns by, additionally  to x.}

\item{facet_rows}{Split rows by}

\item{remove_missing}{If TRUE any missing values in character or factor
columns are remove. If FALSE they are recoded as (Missing).}

\item{label_width}{Width of labels for text wraping}

\item{denom}{Denominator to be used for frequency calculations.
Possible values are "n", "N_col", "N_row".
See also \link{[https://docs.roche.com/doc/tern/v0.7.5/reference/summarize_variables.html]{tern::summarize_vars()}}}
}
\description{
Summry table
Table to display data that is being used in plots. All character columns
are converted into factors.
}
