% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{theme_nightowl}
\alias{theme_nightowl}
\title{Nightowl ggplot2 theme}
\usage{
theme_nightowl(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\value{
ggplot2 theme object
}
\description{
Nightowl ggplot2 theme
}
